refactor: remove unused CanvasTool, adjust imports for updated project structure
