+++ app/lib/generateEphemeralKey.ts
